# src/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cinema API",
    "version": "1.0.0",
    "description": "Cinema Api in TS"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/shows": {
      "get": {
        "summary": "List all shows",
        "description": "Retrieve a list of shows with pagination and sorting options.",
        "operationId": "listShows",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of shows to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "ascending",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Field to order the shows by",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of shows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Show"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/shows/{id}": {
      "post": {
        "summary": "Create a show",
        "description": "Create a show",
        "operationId": "createShow",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startAt",
            "in": "query",
            "description": "StartAt of the show",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "endAt",
            "in": "query",
            "description": "EndAt of the show",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "state of the show",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "room",
            "in": "query",
            "description": "Room of the show",
            "required": true,
            "schema": {
              "type": "object",
              "default": 10
            }
          },
          {
            "name": "movie",
            "in": "query",
            "description": "Movie of the show",
            "required": true,
            "schema": {
              "type": "object",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A show",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "items": {
                        "$ref": "#/components/schemas/Show"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "List all stats",
        "description": "Retrieve a list of stats with sorting options.",
        "operationId": "listMovies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "movie",
            "in": "query",
            "description": "To filter by movie",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "room",
            "in": "query",
            "description": "To filter by room",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "startAt",
            "in": "query",
            "description": "To filter by startAt",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "endAt",
            "in": "query",
            "description": "To filter by endAt",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Stats"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Show": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          },
          "state": {
            "type": "boolean"
          },
          "room": {
            "type": "object"
          },
          "movie": {
            "type": "object"
          }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "visitsCount": {
            "type": "integer",
            "format": "int64"
          },
          "Shows": {
            "type": "object"
          },
          "mostPopularShow": {
            "type": "object"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}